feat: render my info form more declaratively
